+++ src/test/java/org/realityforge/replicant/client/gwt/TestWebPollerDataLoaderService.java
@@ -44,13 +44,6 @@ final class TestWebPollerDataLoaderService
     return TestGraph.class;
   }
 
-  @Nonnull
-  @Override
-  protected String deriveDefaultURL()
-  {
-    return "";
-  }
-
   @Override
   protected void setupCloseHandler()
   {